Add thread-local variable indicating connection currently being handled.
To enable diagnostics in the low-level code to indicate which connection
is responsible for the current operation without having to pass conn*
absolutely everywhere, set a thread-local variable when we enter the
state machine processing.

Accessing a thread-local variable is slower (and more awkward) than
accessing a parameter, so this isn't intended to replace the conn*
parameter in places where the conn data is used in the normal operation
of the code.